+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## V1.45
+### script
+* BatteryMode: save latest five PanelMinVoltages and return the highest value of them. This ignores temporarly DTU-Errors (e.g. reset values at midnight) for maximum of five iterations.
+
+## V1.44
+### script
+* replaced the feature "jump to max limit" to "jump to defined limit"
+### Config
+* changed `[COMMON]/JUMP_TO_MAX_LIMIT_ON_GRID_USAGE` to `[COMMON]/ON_GRID_USAGE_JUMP_TO_LIMIT_PERCENT`
+
 ## V1.43
 ### script
 * Bugfix: timeout OpenDTU